# Conflicts:
#	pom.xml
#	src/main/java/com/bibliomanager/library/controller/AuthorController.java
#	src/main/java/com/bibliomanager/library/controller/BookController.java
#	src/main/java/com/bibliomanager/library/model/Editor.java
#	src/main/java/com/bibliomanager/library/repository/GenreRepository.java
#	src/main/java/com/bibliomanager/library/repository/TypeRepository.java
#	src/main/java/com/bibliomanager/library/service/AuthorService.java
#	src/main/java/com/bibliomanager/library/service/BookService.java
#	src/main/resources/dataLibrary.db

+++ src/main/java/com/bibliomanager/library/repository/EditorRepository.java
@@ -13,7 +13,7 @@ import java.util.List;
  @Repository: repository in the persistence layer and makes it eligible for Spring’s exception translation mechanism.
 
 */
-public interface EditorRepository extends CrudRepository<Editor, Integer> {
+public interface EditorRepository extends CrudRepository<Editor, Long> {
 
     @Query("SELECT COUNT(e) FROM Editor e")
     long countEditors();
